mock_psu: Consolidate components into one file
This commit consolidates all the individual Mock PSU C files into a
single C file. This follows the conventions followed by most of the rest
of the standard modules.
