+++ module/mock_psu/src/Makefile
@@ -7,7 +7,6 @@
 
 BS_LIB_NAME := mock_psu
 BS_LIB_SOURCES := \
-    mod_mock_psu_module.c \
-    mod_mock_psu_psu_driver_api.c
+    mod_mock_psu.c
 
 include $(BS_DIR)/lib.mk
