--- module/mock_psu/src/mod_mock_psu_private.h
@@ -1,17 +1,0 @@
-/*
- * Arm SCP/MCP Software
- * Copyright (c) 2017-2019, Arm Limited and Contributors. All rights reserved.
- *
- * SPDX-License-Identifier: BSD-3-Clause
- */
-
-#ifndef MOD_MOCK_PSU_PRIVATE_H
-#define MOD_MOCK_PSU_PRIVATE_H
-
-#include <stdbool.h>
-#include <stdint.h>
-#include <mod_mock_psu_module_private.h>
-#include <mod_mock_psu_psu_driver_api_private.h>
-#include <mod_mock_psu.h>
-
-#endif /* MOD_MOCK_PSU_PRIVATE_H */
