--- module/mock_psu/src/mod_mock_psu_psu_driver_api_private.h
@@ -1,15 +1,0 @@
-/*
- * Arm SCP/MCP Software
- * Copyright (c) 2017-2019, Arm Limited and Contributors. All rights reserved.
- *
- * SPDX-License-Identifier: BSD-3-Clause
- */
-
-#ifndef MOD_MOCK_PSU_PSU_DRIVER_API_PRIVATE_H
-#define MOD_MOCK_PSU_PSU_DRIVER_API_PRIVATE_H
-
-#include <mod_psu.h>
-
-extern const struct mod_psu_driver_api __mod_mock_psu_psu_driver_api;
-
-#endif /* MOD_MOCK_PSU_PSU_DRIVER_API_PRIVATE_H */
